requirements.txt
streamlit
cmake
opencv-python-headless
pandas
numpy
requests
Pillow
pyyaml
streamlit-authenticator
streamlit-cookies-controller
